stock_api/requirements.txt
Django==3.1.4
django-environ==0.10.0
djangorestframework==3.12.2
joblib==1.2.0
scikit-learn==1.2.2
psycopg2-binary